% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathfinder.R
\name{findPath}
\alias{findPath}
\title{Find best subset of points for follow-up experiments}
\usage{
findPath(tp, y, training, numSubSamples, multiple = F, spline = 1,
  resampleTraining = T, iter = 20, knots = 100, numPerts = 1000)
}
\arguments{
\item{tp}{A numerical vector of time points (or spatial coordinates along a single axis)}

\item{y}{A numerical vector of measurements (of the control)-- this can be set to \code{rep(0, length(tp))} for calculating the shape of the curves only}

\item{training}{Unless multiple is true, this is a numerical matrix of training data, where the rows represent different samples, columns represent different time points (or points on a single spatial axis), and the values correspond to measurements.  If multiple is true, this is a list of numerical matrices, each representing a different class of curves (for instance, different genes).}

\item{numSubSamples}{integer that represents the number of time points that will be subsampled}

\item{multiple}{A boolean that designates whether training is a matrix or a list of matrices}

\item{spline}{A positive integer representing the spline used to interpolate between knots when generating perturbations.  Note that this does NOT designate the spline used when calculating the L2-error.}

\item{resampleTraining}{A boolean designating whether the exact training data should be used (False) or whether a probability distribution of curves should be generated and training curves resampled (True).}

\item{knots}{A positive integer-- for time warping to work optimally, the points must be evenly sampled.  This determines how many points do we evenly sample before conducting time warping}

\item{iterations}{A positive integer, representing the maximum number of iterations employed during time warping (see time_warping in fdasrvf library)}

\item{numPert}{a positive integer, representing the number of sampled curves to output.}
}
\value{
An integer vector of the indices of the time points selected to be subsampled.  The actual time points can be found by \code{tp[output]}.  The length of this vector should be \code{numSubSamples}.
}
\description{
findPath is the main function of the NITPicker package-- it finds the best subset of points to sample from a time course (or spatial axis, along a single axis), based on a set of example curves.
}
